#33 streaming API compiles and unit tests * Test drive pending * Removed 'a lifetime from Transport, was it needed? * File transport modified to do headers in JSON only, body in raw (breaking change). Can be made into compatible solution if needed. * Refactored strongly the SmtpClient and InnerClient to be able to adapt to new API * Transport trait reduced to StreamingTransport, extra features are in new Transport. * Split SmtpTransport from smtp_client module into transport module as the file was getting huge
