+++ src/lib.rs
@@ -32,29 +32,91 @@ pub use crate::smtp::client::net::ClientTlsParameters;
 #[cfg(feature = "smtp-transport")]
 pub use crate::smtp::{ClientSecurity, SmtpClient, SmtpTransport};
 
+use async_std::io::{copy, Write};
 use async_trait::async_trait;
 use std::time::Duration;
 
 /// Transport method for emails
 #[async_trait]
-pub trait Transport<'a> {
+pub trait StreamingTransport {
     /// Result type for the transport
-    type Result;
     type StreamResult;
+    /// Get the default timeout for thistransport
+    fn default_timeout(&self) -> Option<Duration> {
+        None
+    }
 
-    /// Start sending e-mail andreturn a stream to write the body to
-    async fn send_stream(
+    /// Start sending e-mail and return a stream to write the body to with a timeout
+    async fn send_stream_with_timeout(
         &mut self,
         email: SendableEmailWithoutBody,
         timeout: Option<&Duration>,
     ) -> Self::StreamResult;
+}
 
+#[async_trait]
+pub trait Transport: StreamingTransport {
+    type SendResult;
     /// Sends the email
-    async fn send(&mut self, email: SendableEmail) -> Self::Result;
-
+    async fn send(&mut self, email: SendableEmail) -> Self::SendResult;
+    // Send an e-mail with a timeout
     async fn send_with_timeout(
         &mut self,
         email: SendableEmail,
         timeout: Option<&Duration>,
-    ) -> Self::Result;
+    ) -> Self::SendResult;
+    /// Start sending e-mail and return a stream to write the body to
+    async fn send_stream(&mut self, email: SendableEmailWithoutBody) -> Self::StreamResult;
+}
+
+#[async_trait]
+pub trait MailStream: Write {
+    type Output;
+    type Error;
+    async fn done(self) -> Result<Self::Output, Self::Error>;
+}
+
+#[async_trait]
+impl<T, S, E> Transport for T
+where
+    T: StreamingTransport<StreamResult = Result<S, E>>,
+    T: Send + Sync,
+    S: MailStream + Unpin + Send + Sync + 'static,
+    S::Error: From<std::io::Error>,
+    S::Error: From<E>,
+    E: 'static,
+{
+    type SendResult = Result<S::Output, S::Error>;
+    /// Sends the email
+    async fn send(&mut self, email: SendableEmail) -> Self::SendResult {
+        self.send_with_timeout(email, self.default_timeout().as_ref())
+            .await
+    }
+    // Send an e-mail with a timeout
+    async fn send_with_timeout(
+        &mut self,
+        email: SendableEmail,
+        timeout: Option<&Duration>,
+    ) -> Self::SendResult {
+        let mut stream = self
+            .send_stream_with_timeout(
+                SendableEmailWithoutBody::new(
+                    email.envelope().clone(),
+                    email.message_id().to_string(),
+                ),
+                timeout,
+            )
+            .await?;
+
+        copy(email.message(), &mut stream).await?;
+        Ok(stream.done().await?)
+    }
+    /// Start sending e-mail and return a stream to write the body to
+    async fn send_stream(
+        &mut self,
+        email: SendableEmailWithoutBody,
+    ) -> <Self as StreamingTransport>::StreamResult {
+        self.send_stream_with_timeout(email, self.default_timeout().as_ref())
+            .await
+    }
 }
